// src/app/models/leads.model.ts
export interface Lead {
    id: number;
    firstName: string;
    lastName: string;
    dob: string;
    email: string;
    phone: string;
    gender: string;
    startBy: string;
    interestedIn: string;
    aim: string;
    source: string;
    status: string;
    createdAt: string;
    createdBy: string;
    updatedAt: string;
    updatedBy: string;
}